+++ src/main/java/org/c4marathon/assignment/common/authentication/model/principal/Principal.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.common.authentication.model;
+package org.c4marathon.assignment.common.authentication.model.principal;
 
 public interface Principal {
 	Object getId();
